fix(store): close a client-visible data leak + 6 logic bugs found in audit
SECURITY — clients could read the whole practice order book. The owner catalog
and orders routes carried no role decorator, RolesGuard passes any route that
declares none, and a client DOES have a workspaceId (jwt.strategy resolves
clients.workspace_id) — so assertWorkspace() let them through. A client could
GET /workspaces/me/products/orders/all and see every order in the practice,
including other clients' names and purchases, plus unpublished draft products.
Verified against production with a throwaway client (200 on both), and 403 on
both after the fix while the owner still gets 200 and the client's own
storefront still works.

Also fixed:
- clearing an optional field was impossible: COALESCE($n, existing) treats
  "clear" and "leave alone" identically, so removing a photo/description/
  compare-at price restored the old value and stock could never go back to
  unlimited. UPDATE is now built from the keys actually present; null clears.
  Verified: omit preserves, null clears.
- a captured payment could still surface as a failure — the belt-and-braces
  verify call ran inside the main try, so its failure showed a payment error
  next to money already taken. Now non-fatal (webhook owns fulfilment).
- oversell race: stock was checked at checkout but only decremented by the
  webhook, so two clients could both take the last unit. Availability and the
  INSERT are now one statement, and in-flight checkouts hold stock (DISTINCT ON
  client, so one client's retries reserve once).
- repeated Buy clicks stacked duplicate orders; an in-flight checkout is now
  resumed instead (same Razorpay order), only while the price is unchanged.
- amount_paise::int overflowed above ₹21.47L order totals -> ::float8 (exact
  for whole paise, far below 2^53).
- no more orphans: checkout fails before writing when Razorpay isn't
  configured, cancels the row if order creation throws, and replaced/removed/
  deleted product photos are now removed from storage.

+++ backend/src/billing/razorpay-webhook.service.ts
@@ -357,8 +357,9 @@ export class RazorpayWebhookService {
                 razorpay_payment_id = $1,
                 razorpay_order_id   = COALESCE(razorpay_order_id, $2)
           WHERE id = $3::uuid AND workspace_id = $4::uuid AND status = 'pending'
+          -- float8, not int: ::int overflows past ₹21.47L order totals.
           RETURNING id, product_id, quantity, product_name,
-                    amount_paise::int AS amount_paise`,
+                    amount_paise::float8 AS amount_paise`,
         paymentId,
         razorpayOrderId,
         productOrderId,
